Enhance path planning configuration and implement A* algorithm support in planner
