+++ helper_functions.py
@@ -9,6 +9,21 @@ def angle_between_points(p1, p2):
 limit_X = 4.5
 limit_Y = 1.9
 
+def extract_order(word):
+    """
+    Extracts the order of characters from the input word.
+    
+    """
+    arr = []
+    for c in word:
+        if c=='T':
+            arr.append(607)
+        elif c=='O':
+            arr.append(606)
+        elif c=='I':
+            arr.append(604)
+    return arr
+
 
 
 def get_speed(steering_degree):
