+++ main.py
@@ -9,12 +9,11 @@ from robotCommands import *
 from conversion import normalize_angle
 from helper_functions import *
 # from shapely.geometry import Polygon  # Ensure this is imported
-from PARAMETERS import *
 from path_algorithms.planner import *
 
 
 word = "OIT"  # Example word to extract order from
-arr=[]
+
 c_pos, c_rot, c_rad = [0,0,0], 0, 0
 t_pos1, t_rot1, t_rad1 = [0,0,0], 0, 0
 t_pos2, t_rot2, t_rad2 = [0,0,0], 0, 0
@@ -22,19 +21,8 @@ t_pos3, t_rot3, t_rad3 = [0,0,0], 0, 0
 t_pos, t_rot, t_rad = [0,0,0], 0, 0
 y_base = [-0.15, 0.28, 0.67]  # List of base positions for the cubes
 z = 1  # Initialize a global variable for iteration count
-y=604
-def extract_order(word):
-    """
-    Extracts the order of characters from the input word.
-    
-    """
-    for c in word:
-        if c=='T':
-            arr.append(607)
-        elif c=='O':
-            arr.append(606)
-        elif c=='I':
-            arr.append(604)
+current_target_id = 604
+
 
 
 def receive_new_desc(desc: DataDescriptions):
@@ -65,7 +53,7 @@ def receive_new_frame(data_frame: DataFrame):
         if ms.id_num == RigidBodyIDs.CAR:
             # Handle the chaser's data
             c_pos, c_rot, c_rad = optitrack_data_handling.handle_frame(ms)
-        if ms.id_num == y:
+        if ms.id_num == current_target_id:
             # Handle the target's data (ctf_cube)
             t_pos, t_rot, t_rad = optitrack_data_handling.handle_frame(ms)
         if ms.id_num == RigidBodyIDs.CUBE_1:
@@ -190,37 +178,7 @@ def GoBack( ):
 
 
 
-# TODO MOVE TO ANOTHER FILE 
-# Function get data where the  robot car and where the cube is and calculate the path to the cube
-def get_path_to_goal(start_pos, goal_pos, cube_obstacles=[]):
-    global z  # Use the global variable z
-    json_file_path = "path_algorithms/map1.json"
-    planning_env = MapEnvironment(json_file=json_file_path)
-    # Initialize the map environment with the JSON file path
-    planning_env.start = np.array([start_pos[0], start_pos[2]])  # Use x and y coordinates for the start position
-    planning_env.goal = np.array([goal_pos[0], goal_pos[2]])  # Use x and y coordinates for the goal position
 
-    # Add dynamic obstacles (e.g., cubes detected in the environment)
-    for cube_pos in cube_obstacles:
-        if cube_pos != goal_pos:  # Skip if cube_pos matches goal_pos
-            print(f"Adding cube obstacle NOT GOAL at position {cube_pos}.")
-            add_cube_obstacle(planning_env, cube_pos)
-
-    # Create an instance of the RCSPlanner with the planning environment
-    print("Creating RRT* planner...")
-    # planner = RRTStarPlanner(planning_env=planning_env, ext_mode='E2', goal_prob=0.40, k=10)
-
-    planner = AStarPlanner(planning_env=planning_env) 
-    print(f"Planning path from {planning_env.start} to {planning_env.goal}...")
-    # Execute the planning algorithm to get the path
-    plan = planner.plan()
-
-    
-    print("Visualizing the map with the computed plan and expanded nodes...")
-    # planner.planning_env.visualize_map(plan=plan, tree_edges=planner.tree.get_edges_as_states(), name='4main'+str(y))  # Convert z to string
-    # print('Successfully planned path')
-    z += 1  # Increment the global variable z
-    return plan
 # use it to go to the base position
 def go_to_goal(goal_pos):
 
@@ -282,7 +240,7 @@ def get_path_to_target():
                 break
          
 try:
-    extract_order(word) 
+    arr = extract_order(word) 
     print(arr)
     y= arr[0]  # Get the first target ID from the array
     send_servo_request(30)
@@ -296,13 +254,12 @@ try:
         #sys.stdout.flush()  # Ensure that the output is flushed immediately
         print("Streaming started. Waiting for data...", flush=True)
         
-        # TODO
-        # GoBack()
+        
        
         #sys.stdout.flush()  # Ensure that the output is flushed immediately
         for i in range(3):
             print(i)
-            y = arr[i]  # Get the current target ID from the array
+            current_target_id = arr[i]  # Get the current target ID from the array
             
             
             out_limits(c_pos, t_pos)
