+++ path_algorithms/planner.py
@@ -2,3 +2,38 @@ from path_algorithms.create_obstacles import add_cube_obstacle  # Import the fun
 from path_algorithms.MapEnvironment import MapEnvironment
 from path_algorithms.RRTStarPlanner import RRTStarPlanner
 from path_algorithms.AStarPlanner import AStarPlanner
+from PARAMETERS import *
+import numpy as np  # Import numpy for array operations
+
+
+# Function get data where the  robot car and where the cube is and calculate the path to the cube
+def get_path_to_goal(start_pos, goal_pos, cube_obstacles=[]):
+    json_file_path = PlannerConfig.MAP_JSON_PATH  # Get the JSON file path from the configuration
+    planning_env = MapEnvironment(json_file=json_file_path)
+    # Initialize the map environment with the JSON file path
+    planning_env.start = np.array([start_pos[0], start_pos[2]])  # Use x and y coordinates for the start position
+    planning_env.goal = np.array([goal_pos[0], goal_pos[2]])  # Use x and y coordinates for the goal position
+
+    # Add dynamic obstacles (e.g., cubes detected in the environment)
+    for cube_pos in cube_obstacles:
+        if cube_pos != goal_pos:  # Skip if cube_pos matches goal_pos
+            print(f"Adding cube obstacle NOT GOAL at position {cube_pos}.")
+            add_cube_obstacle(planning_env, cube_pos)
+
+    # Create an instance of the RCSPlanner with the planning environment
+    print("Creating A*/RRT* planner...")
+    # planner = RRTStarPlanner(planning_env=planning_env, ext_mode='E2', goal_prob=0.40, k=10)
+    if PlannerConfig.ALGORITHM == "RRT_STAR":
+        planner = RRTStarPlanner(planning_env=planning_env, ext_mode=PlannerConfig.EXTENSION_MODE, goal_prob=PlannerConfig.GOAL_PROBABILITY, k=PlannerConfig.K_NEAREST)
+    if PlannerConfig.ALGORITHM == "ASTAR":
+        planner = AStarPlanner(planning_env=planning_env)
+    
+    print(f"Planning path from {planning_env.start} to {planning_env.goal}...")
+    # Execute the planning algorithm to get the path
+    plan = planner.plan()
+
+    
+    print("Visualizing the map with the computed plan and expanded nodes...")
+    # planner.planning_env.visualize_map(plan=plan, tree_edges=planner.tree.get_edges_as_states(), name='4main'+str(y))  # Convert z to string
+    # print('Successfully planned path')
+    return plan
